Extract UserOptions into its own class
