+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/today/UpcomingWidgetConfigurationPanel.java
@@ -84,24 +84,6 @@ public class UpcomingWidgetConfigurationPanel extends LinearLayout {
         this.listener = listener;
     }
 
-    class UserOptions {
-        private final float opacityLevel;
-        private final WidgetVariant widgetVariant;
-
-        UserOptions(float opacityLevel, WidgetVariant widgetVariant) {
-            this.opacityLevel = opacityLevel;
-            this.widgetVariant = widgetVariant;
-        }
-
-        float getOpacityLevel() {
-            return opacityLevel;
-        }
-
-        WidgetVariant getWidgetVariant() {
-            return widgetVariant;
-        }
-    }
-
     interface ConfigurationListener {
 
         ConfigurationListener NO_OP = new ConfigurationListener() {
