+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/today/UpcomingWidgetConfigureActivity.java
@@ -109,7 +109,7 @@ public class UpcomingWidgetConfigureActivity extends ThemedMementoActivity {
     }
 
     private void saveConfigurations() {
-        UpcomingWidgetConfigurationPanel.UserOptions userOptions = configurationPanel.getUserOptions();
+        UserOptions userOptions = configurationPanel.getUserOptions();
         preferences.storeUserOptions(userOptions);
         new TodayPeopleEventsView(this, AppWidgetManager.getInstance(this)).requestUpdate();
     }
